move bind-call-ui to setup/ module as quick fix for boundary violations, in preperation for call refactor
