+++ src/setup/bind-call-ui.js
@@ -1,8 +1,9 @@
-// src/ui/core/bind-call-ui.js
-
-import { onCallConnected, onCallDisconnected } from './call-lifecycle-ui.js';
-import { hideOutgoingCallingUI } from '../../../../features/call/components/outgoing-call.js';
-import { messagesUI } from '../../../../features/messaging/components/messages-ui.js';
+import {
+  onCallConnected,
+  onCallDisconnected,
+} from '../shared/components/ui/core/call-lifecycle-ui.js';
+import { hideOutgoingCallingUI } from '../features/call/components/outgoing-call.js';
+import { messagesUI } from '../features/messaging/components/messages-ui.js';
 
 /** Wires UI-only handlers to CallController. Business logic handlers are in main.js. */
 export function bindCallUI(CallController) {
