+++ src/shared/components/ui/core/call-lifecycle-ui.js
@@ -1,6 +1,6 @@
 // src/ui/core/call-lifecycle-ui.js
 import { uiState } from './ui-state.js';
-import { enterCallMode, exitCallMode } from './legacy/call-mode.js';
+import { enterCallModeUI, exitCallModeUI } from './legacy/call-mode.js';
 
 export function onCallingStarted() {
   uiState.setView('calling');
@@ -14,10 +14,10 @@ export function onCallingEnded() {
 
 export function onCallConnected() {
   uiState.setView('connected');
-  enterCallMode(); // legacy - remove after migration
+  enterCallModeUI(); // legacy - remove after migration
 }
 
 export function onCallDisconnected() {
   uiState.setView('lobby');
-  exitCallMode(); // legacy - remove after migration
+  exitCallModeUI(); // legacy - remove after migration
 }
